+++ app/src/main/java/com/example/mediaplayer/data/models/VideoInfo.kt
@@ -3,12 +3,13 @@ package com.example.mediaplayer.data.models
 import android.net.Uri
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
+import java.time.Duration
 
 @Parcelize
 data class VideoInfo(
     val contentUri: Uri,
     val name:String?,
-    val duration: Int?,
+    val duration: Long?,
     val size: Long?,
     val quality:Int,
     val realPath:String
